refactor: :recycle: Refatorando todos os arquivos de rota
para adicionar a responsabilidade dos controllers sera necessario trocar os antigos controllers para routes
