+++ backend/src/modules/accounts/accounts.module.ts
@@ -4,8 +4,8 @@ import { PassportModule } from '@nestjs/passport';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
 import { EnsureAuthenticate } from '../../shared/infra/http/middlewares/ensureAuthenticate';
-import { ProfileController } from './infra/controllers/profile.controller';
-import { UsersController } from './infra/controllers/users.controller';
+import { ProfileRouter } from './infra/routes/profile.routes';
+import { UsersRouter } from './infra/routes/users.routes';
 import { User } from './infra/typeorm/entities/User';
 import { UsersRepository } from './infra/typeorm/repositories/UsersRepository';
 import { BCryptHashProvider } from './providers/HashProvider/implementations/BCryptHashProvider';
@@ -13,10 +13,10 @@ import { CreateUserUseCase } from './useCases/createUser/createUserUseCase';
 import { ShowProfileUserUseCase } from './useCases/showProfileUser/showProfileUserUseCase';
 import { jwtConstants } from '../../shared/infra/http/middlewares/constants';
 import { NestJWTTokenProvider } from './providers/TokenProvider/implementations/NestJWTTokenProvider';
-import { SessionsController } from './infra/controllers/sessions.controller';
+import { SessionsRouter } from './infra/routes/sessions.routes';
 import { AuthenticateUserUseCase } from './useCases/authenticateUser/authenticateUserUseCase';
 import { ForgotPasswordUseCase } from '../accounts/useCases/forgotPassword/forgotPasswordUseCase';
-import { PasswordController } from './infra/controllers/password.controller';
+import { PasswordRouter } from './infra/routes/password.routes';
 import { UserTokensRepository } from './infra/typeorm/repositories/UserTokensRepository';
 import { MailsModule } from '../../shared/container/providers/MailProvider/mails.module';
 import { ResetPasswordUseCase } from './useCases/resetPassword/resetPasswordUseCase';
@@ -32,10 +32,10 @@ import { CreateUserController } from './useCases/createUser/createUserController
     forwardRef(() => MailsModule),
   ],
   controllers: [
-    ProfileController,
-    UsersController,
-    SessionsController,
-    PasswordController
+    ProfileRouter,
+    UsersRouter,
+    SessionsRouter,
+    PasswordRouter
   ],
   providers: [
     EnsureAuthenticate,
